requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# 日志
loguru>=0.7.0

# 深度学习和模型
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
deepctr-torch>=0.2.9

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# 图像处理和OCR
Pillow>=10.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
python-bidi>=0.4.2
Shapely>=2.0.0
pyclipper>=1.3.0

# HTTP请求
requests>=2.31.0
httpx>=0.24.0

# 工具类
python-multipart>=0.0.6
python-dotenv>=1.0.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0